Java - Yasjl: Remove assumption that default charset is UTF-8
Motivation
==========
When navigating the JSON structure we want to make sure we match
even if the platform default charset is not UTF-8.

In practice the fields we search for contain only ASCII characters.
The default charset on most platforms is compatible with ASCII,
so this hasn't caused problems. But it's a quick and easy fix to be
precise about the expected encoding.

Change-Id: Ic7f5082d3d3df7949fbc74af8ed4b56115c94f06
Reviewed-on: http://review.couchbase.org/107133
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>