Java - Gardening: JsonPointerCallback should accept byte arrays
Motivation
==========
Our callbacks always want to work with byte arrays. We can eliminate
a lot of boilerplate by converting the matching ByteBuf to a byte array
before dispatching to the callback.

Also, use Jackson to decode string values; stripping the
leading and trailing quote is not sufficient when the values
contain escape sequences.

Change-Id: I141e3ff1f3fe09a7a3fef5f17236604cecd5bad9
Reviewed-on: http://review.couchbase.org/107136
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: David Nault <david.nault@couchbase.com>