JCBC-1316, JCBC-1313, JCBC-1315: Adjust blocking semantics of rows.
Motivation
----------
This changeset makes sure that for both the async and the blocking API
(note that the reactive API does not change) the semantics are different
now: before returning the result/completing the future, all the rows
are buffered up.

This means if someone wants to use an API where they can control
how much data is buffered from the network they need to use the
reactive API directly.

The main reason being that java 8 streams are problematic with
unsubscriptions, i.e. the limit() operator does not unsubscribe
the stream. This and other operators with similar semantics can
lead to the unpleasant state where we would stop fetching rows
from the IO layer and would never start again, leaving the socket
in a half-finished state and we cannot use it again.

Since this is totally unobvious from using the blocking API, we
resort back to buffering the full data upfront to avoid this
situation. For 99% of the users this will be the right tradeoff,
and if you want to load the full database through query then
you need to use the reactive API for that.

The async API is also covered by this since otherwise the user
might end up blocking on the stream inside the callback which
could also cause to stall and deadlock our IO threads. not good!

Change-Id: I534260e84d80dea57ac7f30c78e588c38ff83722
Reviewed-on: http://review.couchbase.org/107152
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>