JVMCBC-649: Add infrastructure for mixing chunking and non-chunking.
Motivation
----------
In many http based services, we need to mix chunked and non-chunked responses. Usually
this is because the actual data queries may return lots of data, but we also need to send
messages to the same socket which does not require stream parsing (like creating indexes).

Modifications
-------------
This changeset adds the generic infrastructure and implements it for the search service
with one request. More types will likely be added soon, both w.r.t services and messages
for each service.

Change-Id: Id2839f0dd287c4bd2da55224cedc6b4cd11ca237
Reviewed-on: http://review.couchbase.org/107240
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>