JVMCBC-650: Add support to capture traffic if enabled.
Motivation
----------
This changeset adds an optin to the IoConfig to enable traffic capturing
on a per service basis (or all of them if no service is explicitly
provided).

It works very much like the netty logging handler, but emits an event
into the event bus instead. Note that this emits at VERBOSE level, so
the logger will also have to enable TRACE logging for it to show up
in the logs.

Change-Id: I706a6c5ac4038de359c0eb45cac5f349f1b18693
Reviewed-on: http://review.couchbase.org/107373
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Graham Pople <grahampople@gmail.com>