JVMCBC-652, JCBC-1319: FTS Index Management Capabilities
Motivation
----------
This changeset adds basic support for FTS index management. One can load
indexes, make changes (for now, only the name) and create new ones or
replace others.

Note that on purpose right now there is not much you can do with an index
once loaded, since we first need to figure out what the right API needs
to be in place for this. But it at least allows a testing infrastructure
to load an index and perform queries against it.

Note: some changes are still pending (and waiting for an RFC spec) like
error handling mapping to exceptions, exposing certain fields, maybe
also allowing to check for index status.

Change-Id: If510cdce0dd3e87e988bbf93dfedd280b4468771
Reviewed-on: http://review.couchbase.org/107512
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>