SCBC-13: Provide rowsAs[T] rather than allowing each row to be converted SCBC-12: Query improvements to track RFC changes
The current rowsAs means all rows have to be converted into a QueryRow,
and then into a representation the user actually wants.  It's
unnecessary, so access to the raw rows is being removed, in favour
of something closer to the Java model of rowsAs and allRowsAs.

Also only returning rows and meta in QueryResult, rather than the
individual meta fields.  Don't entirely agree with this, but it's what
in SDK3 RFC now, and it is consistent with the reactive API.

All changes have been made for the analytics service, while I'm there.

Change-Id: I28cfab1e73a31d7ec18de0aaefe94c3d49c5cc9c
Reviewed-on: http://review.couchbase.org/107544
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Graham Pople <grahampople@gmail.com>