SCBC-11: Port environment creation from Java core
Motivation
----------
Until now, configuration of the Scala client has required creating a
CoreEnvironment, using Java primitives.  This isn't consistent with the
rest of the Scala client, so this change provides a Scala wrapper for
CoreEnvironment.

Future work
-----------
Unfortunately this means we have e.g. two IoConfigs in scope now.
Perhaps we should make the core interface private, and provide a Java
wrapper similar to this one?  The downside of course would be that the
Java wrapper would basically be a C&P of the core interface.

Some parts will need documenting later.

Changes
-------
Add Scala versions of all CoreEnvironment-related primitives.

In addition, took the opportunity to fix other environment-related
things:

Move the ExecutionContext (thread-pool) from a static field in Cluster,
to a member of ClusterEnvironment.  This ensures that environments do
not share resources, and be shut down independently.

Emulate the OwnedSupplier concept from Java, and fix the shutdown logic
appropriately.

Made some docs changes in Java.

Post offline discussion: removing reactive and async shutdown variants,
from both Java and Scala.  We feel it's adding complexity for little
benefit, and can always be re-added later.

Change-Id: Ia6a62487fb6715f893cb5938ea371c0364edb05e
Reviewed-on: http://review.couchbase.org/107818
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Graham Pople <grahampople@gmail.com>