WIP JVMCBC-617 Populate environment from system properties
Modifications
=============
Make the core environment builder and IoConfig builder fully mutable
(builders all the way down) instead of using immutable child configs.
This makes it possible to alter one of a child config's properties
without having to reconstruct the whole immutable configuration.

Add a SystemPropertyPropertyLoader and apply the system properties
to the builder when the core environment is created (after the user
has customized the builder, so system properties have precedence over
values configured in code).

Add a parser for the Golang duration format. Use it to parse timeout
durations.

Change-Id: Ib153c28adc7900d1c2844a9f5ed5bdfa16dd0e05
Reviewed-on: http://review.couchbase.org/107926
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>