Java & Scala - Add sdk-testcases support, and fix 2 search Scala bugs
sdk-testcases is a repo of JSON responses from the Couchbase services,
that we can use to replicate tricky issues, and simply to allow doing
more unit testing.

I'd like to expand our support for sdk-testcases in the future, but
to reduce scope and get us started this just adds support for one of the
search tests.

Adds SearchMock to Java and Scala, which mocks out the search handling
code completely so that it can be fed JSON files.

Add initial tests to Java and Scala for the alltimeouts.json test case.

Add sdk-testcases as a git submodule under the resources dir of
java-client and scala-client.  There's likely some more work to do here
in terms of putting git submodule init into the Makefile, that kind of
thing, but sdk-testcases will likely be moving soon so I'll do this work
after.

Update: was unable to get buildbot/jenkins working with submodules, after
some efforts.
Trying just adding the files directly.  If this works I'll need to come
back later and have another go.

Moved search code into AsyncCluster.searchQuery, for tidiness.

This actually caught two bugs in Scala search, so fixed them:

1. Throw while parsing SearchQueryRow, rather than returning Try.  This a)
should be faster and b) avoids a bug in FunctionalUtil.traverse, and c)
makes a much nicer interface in SearchResult as the Try doesn't need to
be handled.

2. Fix ordering of failed and successful metrics parameters.

Change-Id: I9af777f3d0c43ea1f048e089627e05842a677123
Reviewed-on: http://review.couchbase.org/111221
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>