JVMCBC-659: Allow to disable native transports via env.
Motivation
----------
Since it might always be possible that netty's transports (both
kqueue/nio as well as boringssl) have a bug it should be possible
to disable them through the environment. In this case the code
will automatically fall-back to the regular java-based transports
which might be slower but are always the safe option.

Change-Id: Ia7d564fdc07652fc289763fbff89d191427c5ae4
Reviewed-on: http://review.couchbase.org/108348
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>