JVMCBC-660: Add support for cluster capabilities.
In Mad Hatter, the server adds a new cluster capabilities section
to the config which exposes a method to negotiate cluster-wide
capabilities.

In the first installment, we are recognizing enhanced prepared statement
support for query, but the code has been set-up in a way that it will
pick anything for any service in the future as long as it is known
in the enum.

Change-Id: I04bd671facfc9cb600f010f0ed898e267c1fcca0
Reviewed-on: http://review.couchbase.org/108440
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>