JVMCBC-664: Align timeout deadline with RFC.
This changeset makes sure that, aligning with the RFC, the deadline
sent to the server is 90% of the user provided timeout. Also, the RFC
states that a floor of 1500ms must be applied so no lower timeouts
are possible.

This changeset makes the modification accordingly to the algorithm
and adds two unit tests which verify the custom timeout policy as
well as the floor enforcement.

Also, this changeset removes the RequestUtil class since it contains
code to adjust the timeout as well which would lead to double adjusting.
By having it in a central place now there is less chance of forgetting
to add it to a new command as well.

Change-Id: I8bca6bb970738983c0ae4e06a1d7dbf6624202ad
Reviewed-on: http://review.couchbase.org/108627
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>