JVMCBC-658: Also replace $HOST on NMVB payloads.
Motivation
----------
It was found that when the client receives a NMVB response with a config payload, it did
not properly replace the $HOST variable with the dispatched node. To help with this in
the future, the replace code has been refactored into the context as well (since the
upcoming http refresher will have to do the same).

Change-Id: I2a41d7f9cdadf44c92723cceb234d6c62f94e956
Reviewed-on: http://review.couchbase.org/108950
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>