JVMCBC-658: Identify node not only based on hostname
Since one can run more couchbase nodes on the same node, it is not
enough to identify a server only by the hostname. As a result, we
introduce a NodeIdentifier that consists of the tuple:

<hostname, manager port>

and use it across the core. Unfortunately this introduces lots of
churn, but we have to pass the identifier around in many places
compared to the hostname alone previously.

Change-Id: If42b6ba446d7705b49d12b9728a74eeef6a37030
Reviewed-on: http://review.couchbase.org/108961
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>