core: re-run reconfig if new config arrived
Motivation
----------
We use an atomic boolean to not run multiple reconfigurations in
parallel to make sure they don't race each other. As a result, we
also need to make sure that if a newer config arrives while a
reconfiguration is in progress, once it is completed we trigger
another one so that really every change is picked up eventually.

Change-Id: I2e2a2f343598ddc4da575ee3280f02d7a9ed927f
Reviewed-on: http://review.couchbase.org/109009
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>