JVMCBC-669: Make sure to recognize removed node during rebalance.
Motivation
----------
During initial rebalance testing, it was found that a removed node
sometimes was not picked up properly and the corresponding instance
removed from the node set.

Modifications
-------------
Root cause of this issue was a logic error which on the one hand
only removed services from nodes that are still around but on the
other hand before completely removing a node only if all services
have been removed.

As a result, if a node got removed "in one shot" services were
still around and the node would not get removed as a result.

Change-Id: I4f6513b9a277a63c85c6db97ca30905ae920fac8
Reviewed-on: http://review.couchbase.org/109033
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>