core: Add request type to RequestNotRetriedEvent.
Motivation
----------
During testing it was found that when this event is raised, there is no
way to figure out out of context from which request type it came from.

As a result to ease debugging from logs only, the event now also prints
the type of request that got cancelled alongside any other metadata.

Change-Id: I55a2b1c16318cbce8eb64af75baf5f771ba87af4
Reviewed-on: http://review.couchbase.org/109119
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>