core: improve bucket open/closed logging.
Motivation
----------
To let users know what's going on, we need to print bucket opened
and closed events, as well as letting them know that a shutdown
has completed successfully.

Note that this only covers the "golden" paths for now, there will
be more changes coming which cover all the different error cases
with appropriate events.

Change-Id: If107650fc948b2ca12b868d8abcf43c255202ab7
Reviewed-on: http://review.couchbase.org/109122
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>