JCBC-1345: Prepared Statement Support.
Motivation
----------
The SDK3 implementation has been lacking prepared statement support
completely, so this changeset adds support and tests for it accordingly.

Modifications
-------------
In addition to supporting the "old" approach with the full encoded plans,
this implementation supports the new approach with the name only, in line
with the SDK-RFC. Depending on if enhanced prepared statements are
available in the cluster config, the client decides which approach to
use and also supports the rolling upgrade scenario.

Note that the missing pieces (error handling and unit tests) have been
filed as follow-up tickets.

Change-Id: Icb3d15a7f223c364c2e47451f5c47da37e9092e4
Reviewed-on: http://review.couchbase.org/109158
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>