JVMCBC-670: EnumSet copyOf trips on null values.
When the server has unknown cluster capabilities as part of the config,
jackson encodes them as null. EnumSet.copyOf has problems with that and
as a result we cannot parse those configs and give up.

To fix this, null values are filtered out beforehand so they cannot
be a problem while performing the EnumSet copy.

Change-Id: I6aa8da282c6f8209d66751903f62ffd680b70a3c
Reviewed-on: http://review.couchbase.org/109259
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>