JVMCBC-671: Avoid accidential pipelining for streaming results.
Motivation
----------
While implementing prepared statements it was found that accidential pipelining
can happen which will lead to results never being completed (cleanup and write
of a new request race each other, leading to invalid state inside the handler).

Modifications
-------------
With this change, we go back to the same model that SDK 2 currently has in place.
Instead of listening on the response completion, the handler will trigger a
completion event to the endpoint once it is fully complete. This is an important
difference since after the initial future completion many streaming rows can
arrive which will increase the delta time between future completion and
actual completion of the full streaming.

Added a unit test to verify this behavior. Unfortunately it touches quite a
couple of IO classes to propagate the endpoint into them, so it is a little
more noisy than I had hoped initially.

Finally, since I had to shuffle some bits around in the endpoint I took the
opportunity to now only attach the listener callback if circuit breakers
are enabled - otherwise there is no need for it anymore.

Change-Id: Iab8d233fb2bfcfdc65793d719860e5b8f5b66cea
Reviewed-on: http://review.couchbase.org/109313
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>