JVMCBC-676: Readd multi kv endpoints support with the right strategy.
Motivation
----------
SDK 2 had this, and it just hasn't been implemented for SDK 3 yet. It
works very similar to SDK 2: we look at the partition id for each key
and then route it with a modulo operation to an index in the endpoint
list. Everything with no partition set will go to index 0, which should
be fine.

Unit and integration test added, mostly ported from the SDK 2 suite
as well.

Change-Id: I76123659a7413c0392c7aebe0a552441565c4b2b
Reviewed-on: http://review.couchbase.org/109660
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>