JVMCBC-668: Refactor code in preparation for collection refresh.
Motivation
----------
The client needs to support refreshing collection IDs in the background
per RFC. To remove the burden from each binding and move it into the core
layer (where it should be anyways), some refactoring is needed.

Modifications
-------------
This changeset does not actually do the refrehs, but refactors the core
APIs and how the bindings use the code: there is now a CollectionIdentifier
which is a triple of <bucket name, scope name, collection name> which is
used by the clients to signal what should be used.

The core then ist taking these identifiers and compares it to a CollectionMap
which actually maps the identifiers to their current collection IDs. this map
is going to be filled / updated in a later changeset.

NOTE: this means right now collections DO NOT WORK when applying only this change,
since part of the mechanics are removed. As a result, this change should only
be applied with the subequent changes into master so that the functionality
is restored and expanded.

Change-Id: I81ab5ba62b1297d6f1c2d018d0eda3fd9db47041
Reviewed-on: http://review.couchbase.org/109665
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>