JVMCBC-668: Implement automatic collection ID refresh.
Motivation
----------
This changeset builds on the previous refactor and implements the
actual collection ID refresh.

Modifications
-------------
The configuration provider now takes on the responsibility of holding
and refreshing the collection manifest if needed. To improve efficiency,
we now fetch the full manifest and apply it. The fetch is either done
when a collection is opened, or when the KV layer signals that a collection
ID is outdated.

Change-Id: I75a4c9bf8d6501e28310557bb6fa6e8de22c94a2
Reviewed-on: http://review.couchbase.org/109666
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>