JVMCBC-676: Still allow KV to be pipelined.
Motivation
----------
The last changed introduced a perf regression: the isFree() check
without adding explicit pipeline support also disabled pipelining
for KV which was not the intended purpose.

Modifications
-------------
This code brings it back to 2.x standards, basically letting the
endpoint know if it is pipelined and if so it always reports itself
as "free".

Note that this also includes a small optimization along the way. If
pipelined is used, there is no need to perform volatile stores and
loads on the number of outstanding operations since it is only ever
used to know if a service is free or not.

Change-Id: Ib122d613e73b4a5bfba3b0042498ef23a4c43879
Reviewed-on: http://review.couchbase.org/109727
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>