core: simplify kv key encoding.
Motivation
----------
Internally, there is too much code duplication how we properly
encode the kv keys depending on if collections are available or
not. This can be cleaned up and also made more gc efficient.

Modifications
-------------
The base kv request class now handles all the checking if collections
are enabled or not, which DRYs up the individual implementations calls.
Also, in profiling with JFR it can be seen that the Unpooled wrapping
of the key and the array copies with collections produce GC which can
be avoided.

So aside from the cleanup, this change uses the pooled allocator to
create a pooled key which reduces the GC load a little bit. For example
wit 25k ops/s, in a 5 min run this reduced TLAB allocations by around 300 MBs
with a very short key.

Change-Id: I66206ebaed498004bbb31dddc9b010fe71ccb626
Reviewed-on: http://review.couchbase.org/109808
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>