JCBC-1282 Update Error Handling for SDK3 (checkpoint)
Rename these exceptions to match RFC:

  DocumentAlreadyExistsException -> KeyExistsException
  DocumentDoesNotExistException -> KeyNotFoundException

Add `forKey(String key)` static factory methods to these exceptions:

  CASMismatchException
  KeyExistsException
  KeyNotFoundException

and hide the other constructors. Also add an accessor for the key.
The idea is that any potentially useful information in the error
message should be accessible programmatically. Might extend this
to other exceptions later.

Modify CASMismatchException to not implement the "retryable" marker
interface, since an operation that failed due to CAS mismatch is not
retryable.

Remove DocumentConcurrentlyModifiedException (not in the RFC) and
replace it with CASMismatchException.

Change-Id: If827bc1fc910245721d41157dda0da24c10335ee
Reviewed-on: http://review.couchbase.org/110175
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>