JVMCBC-681: Avoid resolving network addresses early
Motivation
----------
This change is very similar to the change made in java sdk 2.7.7, which
makes sure that hostnames are not resolved early (which prevents it from
working against external kubernetes clusters for example).

This change goes a bit further though and completely removes the
NetworkAddress, since in 3.x we do not need to care about backwards
compatibility at this point.

The end result is simpler code which hopefully works under more
setups.

Change-Id: I84c3d6a7542b38817076027941b811eb4d099313
Reviewed-on: http://review.couchbase.org/110251
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>