JVMCBC-686: Fix default collection support when collections enabled.
Motivation
----------
A regression has been introduced in alpha.4 when collections are enabled
but you want to use the default collection. Since for the default collection
at a higher level we are not checking the manifest, the code runs into issues
since the map is not populated.

Modifications
-------------
This changeset changes this by changing up the CollectionMap a little bit where
the default ID is stored and returned if the identifier is a "default" one.

To make this work I had to add the default property to the identifer, and along
the way I found a bug in the provider which prevented a collection ID from being
refreshed, this is also fixed along the way.

Note that while not particularly necessary, I changed the map from inheritance
to composition since it better encapsulates what we support, and in previous
code reviews this was suggested.

Change-Id: Ia49373801e8a1003f2ab74d3a8831d9f1478e7d8
Reviewed-on: http://review.couchbase.org/110712
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>