JVMCBC-688: Make sure cas-based observe fails for concurrent modifications.
Motivation
----------
An optimization put into place failed to hold the semantics we need
for cas-based observe: when we do not send the observe to the master,
even if just a replica is asked for, we cannot detect the concurrent
CAS modifications and it will keep polling.

Modifications
-------------
Always fetch the active observe, even if only called for replication.
A test case has been added which performs multithreaded ops on the
same key to provoke this state.

Change-Id: I919443466ad3a2407c365174f123f1f506bd7442
Reviewed-on: http://review.couchbase.org/111065
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>