JCBC-1379: Remove Optional from Get-* APIs
Motivation
----------
For consistency and usability reasons, we decided to ditch the
Optional<?> for get*-based APIs and if a document is not found
raise a KeyNotFoundException instead.

We are going to introduce an option where, if the overhead of
an exception is too much, it will be possible to instead return
a default value. This will be provided in a follow-up commit.

Change-Id: I48f683150c043e10b42f75393f4060d6ffc8a9fd
Reviewed-on: http://review.couchbase.org/111206
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>