JVMCBC-689: [Subdoc] Getting `com.couchbase.client.core.error.DecodingFailedException` while performing sub_doc operations
Any subdoc mutateIn counter operations currently fail with a
DecodingFailedException.

Turns out the code was, if the body was present - which is only is if
there's any return values, e.g. if a counter op was performed -
assuming that there's one return value for each command.

This fixes that assumption and only reads values that are there.

Change-Id: I063fdfdaeb3e02d0beafa1f374f32fae170ad64c
Reviewed-on: http://review.couchbase.org/111310
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>