JVMCBC-697: Remove logging from copy of reactor retry logic
This is code that we've copied over from project reactor.

The code currently logs directly to a Logger.  This is a) quite heavy
and b) bypasses the event-based logging we've moved to.

For now, let's remove this logging.  We may want some event-based
logging on retries later, and we'll add it as and when we see where it's
required.

Change-Id: I6d6078a7257ed711da532804d08e7d29e7dddac1
Reviewed-on: http://review.couchbase.org/111368
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>