JVMCBC-696: Make each component in the state tree stateful.
Motivation
----------
This changeset is a preparation for upcoming changes, but it definitely
serves a big purpose on its own: it brings back the idea that a component
is a composite state based on its child components. We had this concept
in the 2.x series and because of lack of time it was never implemented
in 3.x.

This is a very important property, because it allows higher level components
to accurately reflect their state based on their child components.

Modifications
-------------
Based on this, this changeset accomplishes the following:

 - Generic stateful components are re-introduced, one for single states and
   one for composite states.
 - The Endpoint is a single stateful component, whith its state basically
   reflecting the one of the underlying channel.
 - The Service is a composite of all its underlying endpoint states.
 - The Node is a composite of all its underlying services.
 - Each component raises events into the event bus at debug level when its
   state changes.

Unit tests have been added (and some had to be modified) to reflect these
changes.

Change-Id: Ib8736f8f36609be3a7e53cd35f3f3692bd4aa541
Reviewed-on: http://review.couchbase.org/111412
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>