JVMCBC-699: Allow a kv endpoint not to be bound to a bucket.
Motivation
----------
This changeset, while touching a lot of files as a fallout unfortunately,
only has one purpose: to make the bucket optional when the endpoint/service
is created.

If it is optional, the select bucket command is omitted, allowing (in the future)
to fetch cluster level configurations from a kv endpoint.

Note that the functionality itself has not been changed at all for now, this
change only allows for the possibility of not sending a bucket name in the
future.

Modifications
-------------
This changeset touches a lot of files because I had to make the bucket in the
channel context optional and so some purely type-related changes came in as
well.

Other than that, tests have been modified and the "optional" is propagated
to the upper layers as well.

Change-Id: I1342159afad028c9f8da4d15a0566230c698e44c
Reviewed-on: http://review.couchbase.org/111416
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>