JVMCBC-700: Print environment info on SDK init
Motivation
----------
In SDK 2 it has proven very useful to print environment-related
diagnostic information at INFO level on startup. We want to bring
this forward to SDK 3 as well.

Modifications
-------------
This changeset adapts the build process to include the version in
the MANIFEST.mf files for core, java and scala. It also modifies
the environment and its underlying config classes to export their
properties in a map format.

When created, the Core creates a CoreCreatedEvent which will dump
the env-related diagnostics at INFO level.

Change-Id: Ib052ecee1559478fe333df5b558b1fe05cd14513
Reviewed-on: http://review.couchbase.org/111546
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>