JVMCBC-694: Add infrastructure to load a global config.
Motivation
----------
Building on the previous global config changeset, this one adds a
"global loader", which is very similar to the carrier loader but
performs it without being attached to a bucket.

Modifications
-------------
To achieve this, certain modifications had to be made:

 - a new request and response introduced to fetch the config
 - a proposed config context for this global config added
 - the node had to be modified to honor a "empty bucket" state, which is
   now namespaced internally under the _$BUCKET_GLOBAL$_ namespace in
   the hashmap.

A integration test has been added which only runs with 6.5 and later, so
a capability to the test framework has been added accordingly.

Change-Id: I2b23da1654b8620ad023a9e2851ca6f37ca469b5
Reviewed-on: http://review.couchbase.org/111648
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>