JVMCBC-694: Add infrastructure to refresh a global config.
Motivation
----------
This changeset adds the GlobalRefresher which, when activated, polls
the global config for changes and advertises them to the configuration
provider.

Modifications
-------------
Note that only the necessary pieces to propose a config to the config
provider has been implemented in this changeset to keep it contained.
A follow-up change will add the plumbing to put it all together.

An integration test has been added which monitors the advertised configs
to make sure once started they are being sent to the config provider.

Change-Id: I614f9c074bf748e5a3819bd71090d4d0f4794da9
Reviewed-on: http://review.couchbase.org/111677
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>