JVMCBC-694: Add plumbing for cluster level queries.
Motivation
----------
This changeset glues all the pieces together to load a global config
if possible and then maintain it to allow cluster level queries without
opening a bucket.

Modifications
-------------
This changeset performs a couple of important tasks.

Most importantly, the Core is modified to take a global config into
account when performing the reconfiguration. If the cluster reports
back with a "no bucket selected" during bootstrap, we know it's an
old cluster and allow it to continue, all other errors for now
are propagated to the caller.

Change-Id: I757bef08c6b04cab34895cb0a184c7245c154cdb
Reviewed-on: http://review.couchbase.org/111686
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>