JVMCBC-704: Keep SeedNodes fresh with config changes.
Motivation
----------
Usually buckets are opened all at once, but we should still make sure
that the seed node list is kept up-to-date once we fetch new configs
should buckets be opened at a later point and the topology changes.

Modifications
-------------
The config provider now only initially looks into the provided seed
nodes and then caches them locally. With every global or bucket level
config change, the seed node list is updated with the most up-to-date
node and port list.

Note that this code does not handle alternate addresses for now, which
will be added in a later commit once we officially introduce that
functionality.

Change-Id: Ieef9aa95b953ed9c440a80ec8b39774271ea40a2
Reviewed-on: http://review.couchbase.org/111695
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>