core: refactor refresher to also propose configs.
Motivation
----------
For consistency reasons, it makes sense that all components call the
propose* methods on the config provider if they have something to propose.
Right now the config provider streams from the refreshers, which is
inconsistent with the loaders and the global refresher.

Modifications
-------------
This changeset makes a "simple" change so that the refreshers propose
the config directly to the config provider rather than push them into
a flux for external consumption.

An integration test had to be modified to introspect the proposed configs
since it was previously looking into the flux which is not present
anymore.

Change-Id: I44e4907222c4c53572def177d94f12631cf9fc3e
Reviewed-on: http://review.couchbase.org/111748
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>