gardening: silence IJ warning on java 9 CRC32 api.
Motivation
----------
While the update(byte[]) method has been present for a long time,
in java 9 it got refactored as a default method into the checksum
interface. Now IJ complains that we are using an API which is present
in Java 9 only if you use such a JDK in the IDE.

Modifications
-------------
While this is totally harmless, there is aan explicit overload
that the default method calls which we can call directly with very
little effort and silence this warning.

(Note Michael N: usually I don't like making changes to satisfy some
IDE constraints, but this one seems like an ok-ish change to make).

Change-Id: I3959c80d95d5d7b075604e855887d1797444f0ab
Reviewed-on: http://review.couchbase.org/111781
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>