JVMCBC-628: Implement the ClusterManagerRefresher
Motivation
----------
This changeset implements streaming-based config refreshing through
the cluster manager, which is used for memcached buckets where kv-based
config polling is not available.

Modifications
-------------
This changeset refactors a couple of other pieces as well:

 - The ManagerLocator is not needed, the "targeted request" logic has been
   moved into the general purpose round robin locator so it can be utilized
   for other services in the future too.
 - The refresher itself builds on the previously added http streaming
   functionality and resubscribes if the stream completes or fails until
   it is explicitly unregistered by the parent provider.

Integration and unit tests added to verify both the golden path and some
edge cases around config completion or failure.

Note that this commit is light on logging/events, these will be added in
a follow-up commit since the KV refresher also needs more of them.

Change-Id: If7ec4a3dc21c793186cf5806098342e2d46ef32a
Reviewed-on: http://review.couchbase.org/111795
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>