JVMCBC-628: Add refresher event logging.
Motivation
----------
The bucket refresh logic can (temporarily) fail at various stages,
so this code adds proper events to diagnose the issues in the field. Note
that most of these errors are not considered fatal so they are logged
at either INFO or DEBUG level depending on the type of event.

This changeset also contains a small refactoring in the cluster manager
refresher so it is easier to grasp what's going on.

Change-Id: I448f8d3a4b9a83504952fb73553d9e77c26bc7cf
Reviewed-on: http://review.couchbase.org/111806
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>